// src/execution/domain/models/NodeType.ts
export enum NodeType {
  GenerateScript = 'generate_script',
  GenerateImagePrompt = 'generate_image_prompt',
  GenerateImage = 'generate_image',
  AnimateImage = 'animate_image',
  GenerateVoice = 'generate_voice',
  GenerateMusic = 'generate_music',
  AssembleVideo = 'assemble_video',
  UploadAsset = 'upload_asset',
  PublishContent = 'publish_content',
  Notify = 'notify',
  CustomPlugin = 'custom_plugin',
}